Fix - MPAIS-59 Damage Accounts print
The images in the printed file are not similar like masterportal
1. The position of number is shifted, not beautiful.
2. If there are several images with different numbers, only the first
number will be printed to all the images.

The solution of point one is to change the image size of original size
and the solution of point two is to change the style id with image name
and unique id, before the style id is only the image name.