MPVERKEHR-165 bugfix - false equation, info chart
The equation for 'Insgesamt Seit' eq 'Seit Jahresbeginn' + 'Im Vorjahr'
is wrong. The reason: calling updateYear, datasets for full weeks are used.
The consequence: the first days and the last days of the year are missing
due to ignored partial weeks.

The solution: updateYear is now based on daily data - which also means that
we need a minimum of two years data on a daily basis, which we have.