MPVERKEHR-165 change source for subscribeLastUpdate
Until now the 'lastUpdate' date ('Letzte Aktualisierung') in the gfi was
based on the first date of the phenomenonTime range. This leads to the
impression that the last update was made over half an hour ago, most of the time.

The solution: Instead of using the phenomenonTime we now use the 'resultTime',
which is the timestamp of the last object detected by the sensor.