MPVERKEHR-180 activate Anteil_SV for Verkehrsportal
For the Verkehrsportal we differentiate between Kfz and
"Schwerverkehrsanteil" (SV). As the database does not provide SV data,
we have to prepare the code for both Kfz and Kfz+SV.

This has to be done in code. Until the database is ready, we can search
for "trafficCountSVAktivierung" to find all code positions where to
uncomment passages to activate Kfz+SV. We will remove those comments
when the database is ready.