Fix - MPVERKEHR-184 the right text before the count
Because the feature data structure in sensor.js was changed and also there are
no SV-Anteil data, now it will show the prefix text the same with Anzahl
or Count.

Later after the SV-Anteil data come, the original prefix texte will be
changed. For Rad, Anzahl: Number and for KFZ, KFZ abs.: Nnumer and
SV-Anteil: Prozent%