MPVERKEHR-190 - fix falsches Jahr im Diagram Legend
The date data is an array, before the first element of array will be used to check in which year, but for years, the first element belongs to the year before.

So now it will add 3 days to the first element for year calendar to make
sure the right year.