bugfix MPVERKEHR-191 subscribeLastUpdate
The time of the last update is not shown in the GFI of verkehrsportal, sometimes.

This is a problem of order. The footer might be late in rendering due to
network traffic. If the info tab performes its tasks faster than the footer,
a mqtt subscription is made without retained handling and recycled for other
subscriptions. As the footer needs retained handling, it is left in the dark.

The solution is to not recycle mqtt subscriptions. This is possible because all
handlers related to a subscribed topic watches out for the packet.retain flag.