MPAIS-96 Fix: show right position of account for the feature of damage accounts which are from tactical zeichen
1. The damage accounts option from tactical tool is a combination of
point (image) and text, up to now the both file import function don not
have the sufficient function to parse the complex features so that the
position of text will always be shifted.

2. The old exported ais damage account file has no problem because the
feature is not a combination of image and text, but only the image, for
example, if user inputs a number 99, a image damage_account_99.jpg will
be automatically set as feature.

3. The solution is similar as parsing the features from old ais exported
file: Parsing the exported damage accounts kml file to replace the
combination with only point feature (old image).