BG 1456 Bugfix: no changes after more gfi themes switches in a window
The problem is: When two different layers are in the same geometry,
after openning the gfi window and switching from different layers, the
data will be same as before.

The reason is: Up to now we watch if thingId is changed to decide to
render new component, but if different layers are in the same geometry,
the thingId is the same and there are no rendering.

But the MeansOfTransport is different, to solve the problem is that the
MeansOfTransport is also watched and if it is changed, new rendering
will be done to get the new data.