BG-1551 fix trafficCount diagram mousehover
Addon TrafficCount: When hovering over a point in a diagram (days,
weeks, years), the Tooltip shows always the first picked date.

The reason: For ChartJS labels only the first dataset is taken. For
the x axis only the first is necessary. But for the Tooltip the date
of the point is needed, but only the label of the x axis is given.

The solution: Add "datetime" to the tooltipItem to use as reference
when hovering.