G31DEV1-162 Change: colors of chart
According to the requirement of schulbehörde, the colors as reference
will be implented in chart diagram.

For Pie, the standard color stays, the colors from schulbehörde will be
in seperate GFI theme integrated, For Line, the colors from schulbehörde
will be directly integrated in complexType function. For Bar, the
default color and mouseover color will be changed into the color from
schulbehörde.