G31DEV1-347-rba-sozialraum-round-values
RBA Fix: Values of GFI "Balkendiagramm" for "sozialraum" are not
rounded correctly.

Source of this problem is the check for the complexType of the feature
rather than the complexType of the sub value. If a complexType of the
feature has no values (e.g. due to privacy issues) the complexType is
invalid and can't be analysed regarding its metadata. So all sub values
are invalid. This is fixed now.